refactor: move wallet helpers from preload to localWallets.js
