+++ electron/preload.js
@@ -3,22 +3,15 @@
  * `/dist/preload.js` using webpack.
  */
 import { ipcRenderer, remote, shell } from 'electron'
-import { join } from 'path'
-import { promisify } from 'util'
-import assert from 'assert'
 import url from 'url'
-import rimraf from 'rimraf'
-import isSubDir from '@zap/utils/isSubDir'
 import fileExists from '@zap/utils/fileExists'
 import dirExists from '@zap/utils/dirExists'
-import { getAllLocalWallets } from '@zap/utils/localWallets'
+import { getAllLocalWallets, getWalletDir, deleteLocalWallet } from '@zap/utils/localWallets'
 import lndBinaryPath from '@zap/utils/lndBinaryPath'
 import lndGrpcProtoDir from '@zap/utils/lndGrpcProtoDir'
 import validateHost from '@zap/utils/validateHost'
 import splitHostname from '@zap/utils/splitHostname'
 import LndConfig from '@zap/utils/lndConfig'
-
-const fsRimraf = promisify(rimraf)
 
 /**
  * List of domains that we will allow users to be redirected to.
@@ -103,54 +96,6 @@ async function generateLndConfigFromWallet(wallet) {
   return Object.assign({}, lndConfig)
 }
 
-/**
- * Returns specified wallet files location
- *
- * @param {*} chain
- * @param {*} network
- * @param {*} wallet
- */
-function getWalletDir(chain, network, wallet) {
-  return join(remote.app.getPath('userData'), 'lnd', chain, network, wallet)
-}
-
-/**
- * Delete a local wallet from the filesystem.
- *
- * @param {object} location - wallet location desc
- * @param {string} location.chain - chain
- * @param {string} location.network - network
- * @param {string} location.wallet - wallet id
- * @param {string} location.dir - Direct location
- * Must either specify @dir or @chain and @network and @wallet
- */
-async function deleteLocalWallet({ chain, network, wallet, dir }) {
-  // returns wallet location based on arguments configuration
-  const getDir = () => {
-    if (typeof dir === 'string') {
-      return dir
-    }
-
-    try {
-      assert(chain && network && wallet)
-      return getWalletDir(chain, network, wallet)
-    } catch (err) {
-      throw new Error(`Unknown wallet: (chain: ${chain}, network: ${network}, wallet: ${wallet}`)
-    }
-  }
-
-  try {
-    const walletDir = getDir()
-    // for security considerations make sure dir we are removing is actually a wallet dir
-    if (!isSubDir(join(remote.app.getPath('userData'), 'lnd'), walletDir)) {
-      throw new Error('Invalid directory specified')
-    }
-    return fsRimraf(walletDir, { disableGlob: true })
-  } catch (e) {
-    throw new Error(`There was a problem deleting wallet: ${e.message}`)
-  }
-}
-
 function getUserDataDir() {
   return remote.app.getPath('userData')
 }
@@ -172,8 +117,8 @@ window.Zap = {
   openHelpPage,
   openTestnetFaucet,
   generateLndConfigFromWallet,
-  getWalletDir,
   getAllLocalWallets,
+  getWalletDir,
   deleteLocalWallet,
   getUserDataDir,
   validateHost,
